# swagger.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "RetailConnect",
                 "version":  "1.0"
             },
    "paths":  {
...
                  "/api/segments":  {
                                        "post":  {
                                                     "tags":  [
                                                                  "Segment"
                                                              ],
                                                     "requestBody":  {
                                                                         "content":  {
                                                                                         "application/json":  {
                                                                                                                  "schema":  {
                                                                                                                                 "$ref":  "#/components/schemas/CreateSegmentRequest"
                                                                                                                             }
                                                                                                              },
                                                                                         "text/json":  {
                                                                                                           "schema":  {
                                                                                                                          "$ref":  "#/components/schemas/CreateSegmentRequest"
                                                                                                                      }
                                                                                                       },
                                                                                         "application/*+json":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/CreateSegmentRequest"
                                                                                                                               }
                                                                                                                }
                                                                                     }
                                                                     },
                                                     "responses":  {
                                                                       "201":  {
                                                                                   "description":  "Created",
                                                                                   "content":  {
                                                                                                   "text/plain":  {
                                                                                                                      "schema":  {
                                                                                                                                     "$ref":  "#/components/schemas/SegmentResponse"
                                                                                                                                 }
                                                                                                                  },
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/SegmentResponse"
                                                                                                                                       }
                                                                                                                        },
                                                                                                   "text/json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/SegmentResponse"
                                                                                                                                }
                                                                                                                 }
                                                                                               }
                                                                               },
                                                                       "400":  {
                                                                                   "description":  "Bad Request",
                                                                                   "content":  {
                                                                                                   "text/plain":  {
                                                                                                                      "schema":  {
                                                                                                                                     "$ref":  "#/components/schemas/ProblemDetails"
                                                                                                                                 }
                                                                                                                  },
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/ProblemDetails"
                                                                                                                                       }
                                                                                                                        },
                                                                                                   "text/json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/ProblemDetails"
                                                                                                                                }
                                                                                                                 }
                                                                                               }
                                                                               }
                                                                   }
                                                 },
                                        "get":  {
                                                    "tags":  [
                                                                 "Segment"
                                                             ],
                                                    "responses":  {
                                                                      "200":  {
                                                                                  "description":  "Success",
                                                                                  "content":  {
                                                                                                  "text/plain":  {
                                                                                                                     "schema":  {
                                                                                                                                    "type":  "array",
                                                                                                                                    "items":  {
                                                                                                                                                  "$ref":  "#/components/schemas/SegmentListItem"
                                                                                                                                              }
                                                                                                                                }
                                                                                                                 },
                                                                                                  "application/json":  {
                                                                                                                           "schema":  {
                                                                                                                                          "type":  "array",
                                                                                                                                          "items":  {
                                                                                                                                                        "$ref":  "#/components/schemas/SegmentListItem"
                                                                                                                                                    }
                                                                                                                                      }
                                                                                                                       },
                                                                                                  "text/json":  {
                                                                                                                    "schema":  {
                                                                                                                                   "type":  "array",
                                                                                                                                   "items":  {
                                                                                                                                                 "$ref":  "#/components/schemas/SegmentListItem"
                                                                                                                                             }
                                                                                                                               }
                                                                                                                }
                                                                                              }
                                                                              }
                                                                  }
                                                }
                                    },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "CreateSegmentRequest":  {
                                                                    "type":  "object",
                                                                    "properties":  {
                                                                                       "name":  {
                                                                                                    "type":  "string",
                                                                                                    "nullable":  true
                                                                                                },
                                                                                       "description":  {
                                                                                                           "type":  "string",
                                                                                                           "nullable":  true
                                                                                                       },
                                                                                       "criteria":  {
                                                                                                        "type":  "string",
                                                                                                        "nullable":  true
                                                                                                    },
                                                                                       "isActive":  {
                                                                                                        "type":  "boolean"
                                                                                                    }
                                                                                   },
                                                                    "additionalProperties":  false
                                                                },
...
                                       "ProblemDetails":  {
                                                              "type":  "object",
                                                              "properties":  {
                                                                                 "type":  {
                                                                                              "type":  "string",
                                                                                              "nullable":  true
                                                                                          },
                                                                                 "title":  {
                                                                                               "type":  "string",
                                                                                               "nullable":  true
                                                                                           },
                                                                                 "status":  {
                                                                                                "type":  "integer",
                                                                                                "format":  "int32",
                                                                                                "nullable":  true
                                                                                            },
                                                                                 "detail":  {
                                                                                                "type":  "string",
                                                                                                "nullable":  true
                                                                                            },
                                                                                 "instance":  {
                                                                                                  "type":  "string",
                                                                                                  "nullable":  true
                                                                                              }
                                                                             },
                                                              "additionalProperties":  {

                                                                                       }
                                                          },
                                       "SegmentListItem":  {
                                                               "type":  "object",
                                                               "properties":  {
                                                                                  "id":  {
                                                                                             "type":  "integer",
                                                                                             "format":  "int32"
                                                                                         },
                                                                                  "name":  {
                                                                                               "type":  "string",
                                                                                               "nullable":  true
                                                                                           },
                                                                                  "isActive":  {
                                                                                                   "type":  "boolean"
                                                                                               },
                                                                                  "memberCount":  {
                                                                                                      "type":  "integer",
                                                                                                      "format":  "int32"
                                                                                                  }
                                                                              },
                                                               "additionalProperties":  false
                                                           },
                                       "SegmentResponse":  {
                                                               "type":  "object",
                                                               "properties":  {
                                                                                  "id":  {
                                                                                             "type":  "integer",
                                                                                             "format":  "int32"
                                                                                         },
                                                                                  "name":  {
                                                                                               "type":  "string",
                                                                                               "nullable":  true
                                                                                           },
                                                                                  "description":  {
                                                                                                      "type":  "string",
                                                                                                      "nullable":  true
                                                                                                  },
                                                                                  "criteria":  {
                                                                                                   "type":  "string",
                                                                                                   "nullable":  true
                                                                                               },
                                                                                  "isActive":  {
                                                                                                   "type":  "boolean"
                                                                                               },
                                                                                  "createdDate":  {
                                                                                                      "type":  "string",
                                                                                                      "format":  "date-time"
                                                                                                  },
                                                                                  "modifiedDate":  {
                                                                                                       "type":  "string",
                                                                                                       "format":  "date-time",
                                                                                                       "nullable":  true
                                                                                                   }
                                                                              },
                                                               "additionalProperties":  false
                                                           },
...
                                   }
                   }
}